Update dimension interface
